requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.13.3
APScheduler==3.10.4
feedparser==6.0.11
loguru==0.7.2
pydantic==2.9.2
httpx==0.27.2
greenlet==3.1.1

# Database drivers
aiosqlite==0.20.0
asyncpg==0.30.0

# Supabase sync
supabase==2.10.0

